Revert "vd_lavc: change auto-probe order to prefer cuda over vdpau-copy"
This reverts commit 5bb41f864d4a244e0df481efb35b14ad8040d56e.

I guess the autoprobing is still too primitive to handle this well. What
it really should be trying is initializing the wrapper decoder, and if
that doesn't work, try another method. This is complicated by hwaccels
initializing in a delayed way, so there is no easy solution yet.

Probably fixes #4865.